requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
audioread==3.0.1
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==43.0.1
cssselect==1.2.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docstring-to-markdown==0.15
et_xmlfile==2.0.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.55.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.2
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
inflection==0.5.1
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
litellm==1.49.6
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
narwhals==1.15.1
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.51.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parsel==1.9.1
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
propcache==0.2.0
Protego==0.3.1
protobuf==5.29.0
pyairtable==3.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
pytz==2024.2
PyYAML==6.0.2
queuelib==1.7.0
redis==5.1.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
Scrapy==2.11.2
seaborn==0.13.2
sentence-transformers==3.3.1
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
spotipy==2.24.0
sqlparse==0.5.3
streamlit==1.40.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tldextract==5.1.2
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.66.5
transformers==4.47.0
Twisted==24.7.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
w3lib==2.2.1
watchdog==6.0.0
Werkzeug==3.0.4
xgboost==2.1.3
yarl==1.15.4
zipp==3.20.2
zope.interface==7.0.3
